# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SlideBanai API",
    "description": "AI-Powered Presentation Platform with Analytics & Collaboration",
    "version": "1.0.0",
    "contact": {
      "name": "SlideBanai Support",
      "url": "https://slidebanai.com/support",
      "email": "support@slidebanai.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://api.slidebanai.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token"
      }
    },
    "schemas": {
      "Presentation": {
        "type": "object",
        "required": ["id", "user_id", "title", "status"],
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "user_id": {
            "type": "integer",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Q1 2026 Product Launch"
          },
          "description": {
            "type": "string",
            "example": "Comprehensive product launch presentation"
          },
          "status": {
            "type": "string",
            "enum": ["draft", "in_progress", "published"],
            "example": "draft"
          },
          "slides_count": {
            "type": "integer",
            "example": 12
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreatePresentationRequest": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": {
            "type": "string",
            "example": "My Presentation"
          },
          "description": {
            "type": "string",
            "example": "A great presentation"
          },
          "status": {
            "type": "string",
            "enum": ["draft", "in_progress", "published"],
            "example": "draft"
          },
          "slides_count": {
            "type": "integer",
            "example": 0
          }
        }
      },
      "Slide": {
        "type": "object",
        "required": ["id", "presentation_id", "slide_number"],
        "properties": {
          "id": {
            "type": "integer"
          },
          "presentation_id": {
            "type": "integer"
          },
          "slide_number": {
            "type": "integer",
            "example": 1
          },
          "content": {
            "type": "string",
            "example": "# Welcome to SlideBanai"
          },
          "background_color": {
            "type": "string",
            "example": "#FFFFFF"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateSlideRequest": {
        "type": "object",
        "required": ["slide_number", "content"],
        "properties": {
          "slide_number": {
            "type": "integer",
            "example": 1
          },
          "content": {
            "type": "string",
            "example": "# Welcome to SlideBanai"
          },
          "background_color": {
            "type": "string",
            "example": "#FFFFFF"
          }
        }
      },
      "Collaborator": {
        "type": "object",
        "required": ["id", "presentation_id", "user_id", "role"],
        "properties": {
          "id": {
            "type": "integer"
          },
          "presentation_id": {
            "type": "integer"
          },
          "user_id": {
            "type": "integer"
          },
          "role": {
            "type": "string",
            "enum": ["owner", "editor", "viewer", "commenter"],
            "example": "editor"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AddCollaboratorRequest": {
        "type": "object",
        "required": ["userId", "role"],
        "properties": {
          "userId": {
            "type": "integer",
            "example": 2
          },
          "role": {
            "type": "string",
            "enum": ["owner", "editor", "viewer", "commenter"],
            "example": "editor"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": ["error", "status"],
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/presentations": {
      "get": {
        "tags": ["Presentations"],
        "summary": "List all presentations",
        "operationId": "listPresentations",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Results per page",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by status",
            "schema": {
              "type": "string",
              "enum": ["draft", "in_progress", "published"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of presentations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Presentation"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Presentations"],
        "summary": "Create a new presentation",
        "operationId": "createPresentation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePresentationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Presentation created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Presentation"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/presentations/{id}": {
      "get": {
        "tags": ["Presentations"],
        "summary": "Get presentation by ID",
        "operationId": "getPresentation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Presentation ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Presentation details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Presentation"
                }
              }
            }
          },
          "404": {
            "description": "Presentation not found"
          }
        }
      },
      "put": {
        "tags": ["Presentations"],
        "summary": "Update presentation",
        "operationId": "updatePresentation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Presentation ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePresentationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Presentation updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Presentation"
                }
              }
            }
          },
          "404": {
            "description": "Presentation not found"
          }
        }
      },
      "delete": {
        "tags": ["Presentations"],
        "summary": "Delete presentation",
        "operationId": "deletePresentation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Presentation ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Presentation deleted"
          },
          "404": {
            "description": "Presentation not found"
          }
        }
      }
    },
    "/presentations/{presentationId}/slides": {
      "get": {
        "tags": ["Slides"],
        "summary": "Get all slides for a presentation",
        "operationId": "listSlides",
        "parameters": [
          {
            "name": "presentationId",
            "in": "path",
            "required": true,
            "description": "Presentation ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of slides",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Slide"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Slides"],
        "summary": "Create a new slide",
        "operationId": "createSlide",
        "parameters": [
          {
            "name": "presentationId",
            "in": "path",
            "required": true,
            "description": "Presentation ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSlideRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Slide created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Slide"
                }
              }
            }
          }
        }
      }
    },
    "/presentations/{presentationId}/slides/{slideId}": {
      "get": {
        "tags": ["Slides"],
        "summary": "Get a specific slide",
        "operationId": "getSlide",
        "parameters": [
          {
            "name": "presentationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "slideId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Slide details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Slide"
                }
              }
            }
          },
          "404": {
            "description": "Slide not found"
          }
        }
      },
      "put": {
        "tags": ["Slides"],
        "summary": "Update a slide",
        "operationId": "updateSlide",
        "parameters": [
          {
            "name": "presentationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "slideId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSlideRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Slide updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Slide"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Slides"],
        "summary": "Delete a slide",
        "operationId": "deleteSlide",
        "parameters": [
          {
            "name": "presentationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "slideId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Slide deleted"
          }
        }
      }
    },
...
    "/api/collaboration/{presentationId}": {
      "get": {
        "tags": ["Collaborators"],
        "summary": "Get all collaborators",
        "operationId": "getCollaborators",
        "parameters": [
          {
            "name": "presentationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of collaborators",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Collaborator"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Collaborators"],
        "summary": "Add collaborator",
        "operationId": "addCollaborator",
        "parameters": [
          {
            "name": "presentationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCollaboratorRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Collaborator added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Collaborator"
                }
              }
            }
          }
        }
      }
    },
...
  }
}